renamed: components폴더명 변경 / 변경에 따른 경로 수정
