+++ src/pages/employer/_components/JobInfoCard/index.tsx
