+++ src/pages/employer/_components/JobInfoTable/index.tsx
